[Edit] scene 4-1 to 4-6
