+++ app/(interactive)/4-6/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function page() {
+  return (
+    <div className="ml-[87px] mt-[232px]">
+      <p>เรื่องที่อยากทำ</p>
+      <p className="mt-3 text-center">เรื่องที่ยังทำไม่ได้</p>
+    </div>
+  )
+}
