+++ package.json
@@ -15,6 +15,7 @@
     "clsx": "^2.0.0",
     "lucide-react": "^0.292.0",
     "next": "14.0.1",
+    "perfect-freehand": "^1.2.0",
     "react": "^18",
     "react-dom": "^18",
     "tailwind-merge": "^2.0.0",
